Major changes
Made some major changes to the game. Condensed planet scripts into only
Planet and PlanetSpawner. Added halo effect and circle trigger collider
for catching planets. Fixed some scripts to use better programming
practices (private variables with getters and setters).  Wanted to add a
different button for catching and throwing because a lot of the time I
would catch a planet and then throw it right away, but couldn't figure
out where the keybindings were. Also added a PlanetState enum, though I
didn't do much with it (only made it so the planets don't destroy when
they are orbiting). Made it so the planets are destroyed when they hit
the boundaries. Probably more stuff but that's all that I remember.